Subcontractor flow

+++ ProbuildBackend/Controllers/BidsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ProbuildBackend.Models;
 using ProbuildBackend.Models.DTO;
@@ -16,17 +17,15 @@ namespace ProbuildBackend.Controllers
             _context = context;
         }
 
-        // GET: api/Bids
         [HttpGet]
         public async Task<ActionResult<IEnumerable<BidModel>>> GetBids()
         {
             return await _context.Bids
-                .Include(b => b.Job) // Include related Job
-                .Include(b => b.User)    // Include related user
+                .Include(b => b.Job)
+                .Include(b => b.User)
                 .ToListAsync();
         }
 
-        // GET: api/Bids/5
         [HttpGet("{id}")]
         public async Task<ActionResult<BidModel>> GetBid(int id)
         {
@@ -43,31 +42,24 @@ namespace ProbuildBackend.Controllers
             return bid;
         }
 
-        // POST: api/Bids
-        [HttpPost]
-        public async Task<ActionResult<BidModel>> PostBid([FromForm] BidDto bidrequest)
+        [HttpPost("upload")]
+        public async Task<ActionResult<BidModel>> PostPdfBid([FromBody] PdfBidDto bidRequest)
         {
+            var userId = User.FindFirstValue("UserId");
+
+            if (string.IsNullOrEmpty(userId))
+            {
+                return Unauthorized();
+            }
+
             var bid = new BidModel
             {
-                Task = bidrequest.Task,
-                Duration = bidrequest.Duration,
-                JobId = bidrequest.JobId,
-                UserId = bidrequest.UserId
+                JobId = bidRequest.JobId,
+                DocumentUrl = bidRequest.DocumentUrl,
+                UserId = userId,
+                Status = "Submitted",
+                SubmittedAt = DateTime.UtcNow
             };
-
-            if (bidrequest.Quote != null)
-            {
-                // Read the quote file into a byte array
-                using (var memoryStream = new MemoryStream())
-                {
-                    await bidrequest.Quote.CopyToAsync(memoryStream);
-                    bid.Quote = memoryStream.ToArray();
-                }
-            }
-            else
-            {
-                bid.Quote = null;
-            }
 
             _context.Bids.Add(bid);
             await _context.SaveChangesAsync();
@@ -75,7 +67,6 @@ namespace ProbuildBackend.Controllers
             return CreatedAtAction(nameof(GetBid), new { id = bid.Id }, bid);
         }
 
-        // PUT: api/Bids/5
         [HttpPut("{id}")]
         public async Task<IActionResult> PutBid(int id, BidModel bid)
         {
@@ -105,14 +96,24 @@ namespace ProbuildBackend.Controllers
             return NoContent();
         }
 
-        // DELETE: api/Bids/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteBid(int id)
         {
+            var userId = User.FindFirstValue("UserId");
+            if (string.IsNullOrEmpty(userId))
+            {
+                return Unauthorized();
+            }
+
             var bid = await _context.Bids.FindAsync(id);
             if (bid == null)
             {
                 return NotFound();
+            }
+
+            if (bid.UserId != userId)
+            {
+                return Forbid();
             }
 
             _context.Bids.Remove(bid);
@@ -121,6 +122,37 @@ namespace ProbuildBackend.Controllers
             return NoContent();
         }
 
+        [HttpPost("{id}/withdraw")]
+        public async Task<IActionResult> WithdrawBid(int id)
+        {
+            var userId = User.FindFirstValue("UserId");
+            if (string.IsNullOrEmpty(userId))
+            {
+                return Unauthorized();
+            }
+
+            var bid = await _context.Bids.FindAsync(id);
+            if (bid == null)
+            {
+                return NotFound();
+            }
+
+            if (bid.UserId != userId)
+            {
+                return Forbid();
+            }
+
+            if (bid.Status != "Submitted")
+            {
+                return BadRequest("Only submitted bids can be withdrawn.");
+            }
+
+            bid.Status = "Withdrawn";
+            await _context.SaveChangesAsync();
+
+            return Ok(bid);
+        }
+        
         private bool BidExists(int id)
         {
             return _context.Bids.Any(e => e.Id == id);
